Backend/requirements.txt
flask
flask-cors
openai>=1.0.0
google-generativeai
python-dotenv
